moved class into own js file for easier testing
